Set up authentication controller and service. Also added swagger and changed scripts to build
