+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "start_dev": "nodemon",
-    "start_prod": "node dist/index.js",
-    "build": "tsc",
+    "start": "node dist/index.js",
+    "build": "tsc && copyfiles -u 1 src/**/*.yaml dist/",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -25,9 +25,11 @@
     "@types/cookie-parser": "^1.4.8",
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
+    "@types/js-yaml": "^4.0.9",
     "@types/node": "^22.10.2",
     "@types/swagger-jsdoc": "^6.0.4",
     "@types/swagger-ui-express": "^4.1.7",
+    "copyfiles": "^2.4.1",
     "husky": "^9.1.7",
     "nodemon": "^3.1.9",
     "ts-node": "^10.9.2",
@@ -42,6 +44,7 @@
     "express": "^4.21.2",
     "firebase": "^11.1.0",
     "firebase-admin": "^13.0.2",
+    "js-yaml": "^4.1.0",
     "mongoose": "^8.9.2",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1"
