+++ src/boostrap.ts
@@ -3,8 +3,9 @@ import bodyParser from 'body-parser';
 import cors from 'cors';
 import mongoose from 'mongoose';
 import swaggerUi from 'swagger-ui-express';
-import swaggerSpec from './config/swagger-config'; 
+import swaggerDocs from './swagger/swagger-index';
 import patientsRoutes from './routes/patients/patientRoutes'
+import indexRoutes from './routes/indexRoute'
 import indexRoute from './routes/indexRoute'
 import errorHandler from "./middlewares/errorMiddleware"
 import dotenv from "dotenv";
@@ -40,7 +41,8 @@ app.use(cors({ credentials: true }));
 app.use(bodyParser.json());
 app.use(indexRoute)
 app.use("/patient", patientsRoutes);
-app.use('/swagger', swaggerUi.serve, swaggerUi.setup(swaggerSpec));
+app.use("/", indexRoutes);
+app.use('/swagger', swaggerUi.serve, swaggerUi.setup(swaggerDocs));
 
 
 export default app;
