+++ src/swagger/auth/loginSwagger.yaml
@@ -1,0 +1,77 @@
+paths:
+  /login:
+    post:
+      summary: Iniciar sesión de un usuario
+      description: Autentica a un usuario utilizando un token y configura una cookie de sesión.
+      requestBody:
+        required: true
+        content:
+          application/json:
+            schema:
+              type: object
+              required:
+                - token
+              properties:
+                token:
+                  type: string
+                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
+      responses:
+        200:
+          description: Login exitoso.
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  message:
+                    type: string
+                    example: "Login exitoso."
+        400:
+          description: Token no válido o vacío.
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  message:
+                    type: string
+                    example: "Token no valido o vacío."
+        401:
+          description: Credenciales inválidas.
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  message:
+                    type: string
+                    example: "Credenciales inválidas."
+        500:
+          description: Error interno en el servidor.
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  message:
+                    type: string
+                    example: "Error al logguear usuario."
+                  error:
+                    type: string
+                    example: "Error desconocido."
+
+  /logout:
+    post:
+      summary: Cerrar sesión de un usuario
+      description: Borra la cookie de sesión del usuario y finaliza la sesión.
+      responses:
+        200:
+          description: Desconexión exitosa.
+          content:
+            application/json:
+              schema:
+                type: object
+                properties:
+                  message:
+                    type: string
+                    example: "Desconectado correctamente."
