+++ tsconfig.json
@@ -9,7 +9,7 @@
         "skipLibCheck": true,
         "forceConsistentCasingInFileNames": true
     },
-    "include": ["src/**/*.ts"],
+    "include": ["src/**/*.ts", "src/**/*.yaml"],
     "exclude": ["node_modules", "**/*.test.ts"]
 
 }
